backend/requirements.txt
# Core dependencies
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-Login==0.6.3
Werkzeug==2.3.7
SQLAlchemy==2.0.25
PyMySQL==1.1.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
cryptography==41.0.7
gunicorn==21.2.0
marshmallow==3.20.1

# Development dependencies
pytest==7.4.3
pytest-cov==4.1.0
black==23.3.0
flake8==6.0.0
flake8-docstrings==1.7.0
mypy==1.5.1
pre-commit==3.3.3
